@@ -1,23 +1,20 @@
 {
-	"name": "server_client",
-	"version": "1.0.0",
-	"description": "Node.js server and client project",
-	"engines": {
-		"node": "18.17.1",
-		"npm": "9.6.7"
-	},
-	"scripts": {
-		"install": "concurrently \"cd server && npm install\" \"cd client && npm install\"",
-		"build": "cd client && npm run build",
-		"start:server": "cd server && npm start",
-		"start:client": "cd client && npm start",
-		"start": "npm run build && concurrently \"npm run start:server\" \"npm run start:client\"",
-		"lint": "eslint .",
-		"lint:fix": "eslint . --fix"
-	},
-	"devDependencies": {
-		"concurrently": "^7.0.0",
-		"eslint": "^8.51.0",
-		"eslint-plugin-import": "^2.28.1"
-	}
+  "name": "full-stack-server",
+  "version": "1.0.0",
+  "engines": {
+    "node": "20.18.0",
+    "npm": "10.8.2"
+  },
+  "scripts": {
+    "setup": "npm install && concurrently \"npm install --prefix server\" \"npm install --prefix client\"",
+    "dev": "concurrently \"npm run server\" \"npm run build\"",
+    "clean": "rm -rf node_modules server/node_modules client/node_modules client/build",
+    "build": "npm run build --prefix client",
+    "client": "npm start --prefix client",
+    "server": "nodemon server/server.js --watch server"
+  },
+  "devDependencies": {
+    "concurrently": "^9.0.1",
+    "nodemon": "^3.1.7"
+  }
 }
